+++ libmicrocosm/process/fork.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
+#include <sys/types.h>
 #include <unistd.h>
 
 #include "reerrno.h"
 
-long MC_fork()
+ssize_t MC_fork()
 {
     pid_t ret;
     REERRNO(ret, fork, -1, ());
