+++ libmicrocosm/time/gettimeofday.c
@@ -1,12 +1,13 @@
 #include <sys/time.h>
+#include <sys/types.h>
 
 #include "conv/struct_timeval.h"
 #include "mcerrno.h"
 #include "reerrno.h"
 
-long MC_gettimeofday(struct MC_struct_timeval *tv, void *tz)
+ssize_t MC_gettimeofday(struct MC_struct_timeval *tv, void *tz)
 {
-    long ret;
+    int ret;
     struct timeval htv;
     if (tz) return -MC_ENOSYS;;
     REERRNO(ret, gettimeofday, -1, (&htv, NULL));
