+++ libmicrocosm/unistd/lseek.c
@@ -4,7 +4,7 @@
 #include "conv/enum_seek.h"
 #include "reerrno.h"
 
-off_t MC_lseek(long fd, off_t offset, long whence)
+off_t MC_lseek(ssize_t fd, off_t offset, ssize_t whence)
 {
     off_t ret;
     REERRNO(ret, lseek, -1, (fd, offset, MC_seek_g2h(whence)));
