+++ libmicrocosm/unistd/write.c
@@ -1,8 +1,9 @@
+#include <sys/types.h>
 #include <unistd.h>
 
 #include "reerrno.h"
 
-ssize_t MC_write(long fd, void *buf, size_t count)
+ssize_t MC_write(ssize_t fd, void *buf, size_t count)
 {
     ssize_t ret;
     REERRNO(ret, write, -1, (fd, buf, count));
